% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDMRs.R
\name{annotateDMRs}
\alias{annotateDMRs}
\title{Annotate DMRs}
\usage{
annotateDMRs(annotation, gff.files, gff3.out, input.dir, out.dir)
}
\arguments{
\item{annotation}{annotation terms used to annotate DMRs}

\item{gff.files}{multiple gff3 files can be supplied as a vector}

\item{gff3.out}{a logical specifying whether output annotated files in gff3 format}

\item{input.dir}{input directory containing filtered DMR matrix/matrices. Ideally any file containing 3 columns i.e (chr, start, stop) can be supplied.}

\item{out.dir}{output directory}
}
\value{
output files containing annotated DMRs and DMR counts table.
}
\description{
This function takes gff3 files as input and outputs annotated DMRs in text and gff3 format. Additionally, a DMR count table is generated.
}
