% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CfromFASTAv4.R
\name{CfromFASTAv4}
\alias{CfromFASTAv4}
\title{Extract cytosines from FASTA}
\usage{
CfromFASTAv4(fasta, chr, out.dir, write.output)
}
\description{
This function slides along the genome and extracts all Cytosines
}
