% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRmatrix.R
\name{makeDMRmatrix}
\alias{makeDMRmatrix}
\title{Builds a DMR matrix for all samples}
\usage{
makeDMRmatrix(
  context = c("CG", "CHG", "CHH"),
  samplefiles,
  input.dir,
  out.dir,
  include.intermediate = FALSE
)
}
\arguments{
\item{context}{cytosine context}

\item{samplefiles}{file containing full path of base level methylation calls, sample names and replicates(optional)}

\item{input.dir}{input directory containing all region level methylome calls}

\item{out.dir}{output directory}

\item{include.intermediate}{A logical specifying whether or not the intermediate component should be included in the HMM model.By default this option is set as FALSE.}
}
\description{
This function generates a binary matrix, a matrix of recalibrated methylation levels and posterior probabilities for all samples.
}
