% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReg.R
\name{makeReg}
\alias{makeReg}
\title{Generates cytosine clusters in genome}
\usage{
makeReg(
  ref.genome,
  contexts,
  chr,
  min.C,
  N.boot,
  N.sim.C,
  fp.rate,
  set.tol,
  out.dir,
  out.name,
  makeRegnull
)
}
\description{
Uses the extracted cytosines from CfromFASTA and finds clsuters of cytosines along the genome.
}
