% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMethimpute.R
\name{runjDMRregions}
\alias{runjDMRregions}
\title{Run jDMR for Regions}
\usage{
runjDMRregions(
  out.dir,
  fasta.file,
  samplefiles,
  genome,
  contexts = c("CG", "CHG", "CHH"),
  include.intermediate = FALSE,
  mincov = 0,
  nCytosines = 0
)
}
\arguments{
\item{out.dir}{output directory}

\item{fasta.file}{path to genome fasta file}

\item{samplefiles}{a text file containing path to samples and sample names. For control/treatment data an additional column specifying the replicates is required.}

\item{genome}{Genome name as a string .e.g Arabidopsis or Human etc}

\item{contexts}{cytosine contexts as a vector. By default this option is set for all 3 cytosine contexts CG, CHG and CHH.}

\item{include.intermediate}{A logical specifying whether or not the intermediate component should be included in the HMM model.By default it is set as FALSE.}

\item{mincov}{Minimum read coverage over cytosines. Default is set as 0.}

\item{nCytosines}{Minimum number of cytosines. Default is set as 0.}
}
\description{
This function runs a HMM model on identified Cytosine clusters
}
