% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMethimpute.R
\name{runMethimputeGrid}
\alias{runMethimputeGrid}
\title{Run Methimpute on binned genome}
\usage{
runMethimputeGrid(
  out.dir,
  fasta.file,
  samplefiles,
  min.C,
  sliding.window = FALSE,
  genome,
  contexts = c("CG", "CHG", "CHH"),
  mincov = 0,
  include.intermediate = FALSE,
  nCytosines = 0
)
}
\arguments{
\item{out.dir}{output directory}

\item{fasta.file}{path to genome fasta file}

\item{samplefiles}{a text file containing path to samples and sample names. For control/treatment data an additional column specifying the replicates is required.}

\item{min.C}{minimum number of cytosines in at least 90 percent of the bins/regions.}

\item{sliding.window}{Logical specifying whether to use sliding windows while binning genome. By default non-sliding window with equal bin-size and step-size are used (set to FALSE).}

\item{genome}{Genome name as a string .e.g Arabidopsis or Human etc}

\item{contexts}{cytosine contexts as a vector. By default this option is set for all 3 cytosine contexts CG, CHG and CHH.}

\item{mincov}{Minimum read coverage over cytosines. Default is set as 0.}

\item{include.intermediate}{A logical specifying whether or not the intermediate component should be included in the HMM model.By default it is set as FALSE.}

\item{nCytosines}{Minimum number of cytosines. Default is set as 0.}
}
\description{
this function runs a HMM model on a genome binned using a sliding/non-sliding window approach
}
