% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMethimpute.R
\name{runMethimputeRegions}
\alias{runMethimputeRegions}
\title{Run Methimpute for Regions}
\usage{
runMethimputeRegions(
  samplefiles,
  Regionfiles,
  genome,
  context,
  out.dir,
  include.intermediate = FALSE,
  mincov = 0,
  nCytosines = 0
)
}
\arguments{
\item{samplefiles}{a text file containing path to samples and sample names, replicate info}

\item{Regionfiles}{output of makeReg, containing coordinates of cytosine clusters/regions}

\item{genome}{genome label for .e.g Arabidopsis}

\item{context}{cytosine context}

\item{out.dir}{output directory}

\item{include.intermediate}{A logical specifying wheter or not the intermediate component should be included in the HMM.By default it is set as FALSE.}

\item{mincov}{Minimum read coverage over cytosines}

\item{nCytosines}{Minimum number of cytsoines}
}
\description{
This function runs a HMM model on identified Cytosine clusters
}
