% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDMRmatrix.R
\name{filterDMRmatrix}
\alias{filterDMRmatrix}
\title{filter DMR matrix}
\usage{
filterDMRmatrix(
  epiMAF.cutoff = NULL,
  replicate.consensus = NULL,
  gridDMR,
  data.dir
)
}
\arguments{
\item{epiMAF.cutoff}{Filter for Minor Epi-Allele frequency. By default this option is set to NULL. Applicable for population level data.}

\item{replicate.consensus}{Percentage of concordance in methylation states among samples with multiple replicates. By default this option is set to NULL. Useful for Control/treatment data}

\item{gridDMR}{Logical specifying if grid DMR approach was used for calling DMRs.}

\item{data.dir}{Directory containing DMR matrix files. Looks for files with suffix, _StateCalls.txt and _rcMethlvl.txt}
}
\description{
Filters non-polymorphic patterns by default.
}
